test: unified timestamp invalid precision
